Issue #SB-2157 feat: Popup implemented with unit testing
