+++ src/app/migration/src/src/app/modules/workspace/components/content-creation-forms/create-course/create-course.component.ts
@@ -1,34 +1,14 @@
-import { Component, OnInit } from '@angular/core';
-import { ResourceService } from '@sunbird/shared';
-import { Router } from '@angular/router';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-create-course',
   templateUrl: './create-course.component.html',
   styleUrls: ['./create-course.component.css']
 })
-export class CreateCourseComponent implements OnInit {
+export class CreateCourseComponent {
 
-  /**
-   * To call resource service which helps to use language constant
-   */
-  public resourceService: ResourceService;
-
-  /**
-  * Constructor to create injected service(s) object
-  *
-  * Default method of DeleteComponent class
-
-  * @param {ResourceService} resourceService Reference of ResourceService
- */
-  constructor(resourceService: ResourceService, private router: Router) {
-    this.resourceService = resourceService;
+  constructor() {
   }
 
-  ngOnInit() {
-  }
-  goToCreate() {
-    this.router.navigate(['/workspace/content/create']);
-  }
 
 }
