+++ src/app/migration/src/src/app/modules/workspace/components/content-creation-forms/create-lesson-plan/create-lesson-plan.component.ts
@@ -1,34 +1,15 @@
-import { Component, OnInit } from '@angular/core';
-import { ResourceService } from '@sunbird/shared';
-import { Router } from '@angular/router';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-create-lesson-plan',
   templateUrl: './create-lesson-plan.component.html',
   styleUrls: ['./create-lesson-plan.component.css']
 })
-export class CreateLessonPlanComponent implements OnInit {
+export class CreateLessonPlanComponent {
 
-  /**
-   * To call resource service which helps to use language constant
-   */
-  public resourceService: ResourceService;
+    constructor() {
+    }
 
-  /**
-  * Constructor to create injected service(s) object
-  *
-  * Default method of DeleteComponent class
 
-  * @param {ResourceService} resourceService Reference of ResourceService
- */
-  constructor(resourceService: ResourceService, private router: Router) {
-    this.resourceService = resourceService;
   }
 
-  ngOnInit() {
-  }
-  goToCreate() {
-    this.router.navigate(['/workspace/content/create']);
-  }
-
-}
