+++ src/app/migration/src/src/app/modules/workspace/components/content-creation-forms/create-study-material/create-study-material.component.ts
@@ -1,34 +1,13 @@
-import { Component, OnInit } from '@angular/core';
-import { ResourceService } from '@sunbird/shared';
-import { Router } from '@angular/router';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-create-study-material',
   templateUrl: './create-study-material.component.html',
   styleUrls: ['./create-study-material.component.css']
 })
-export class CreateStudyMaterialComponent implements OnInit {
+export class CreateStudyMaterialComponent {
 
-  /**
-   * To call resource service which helps to use language constant
-   */
-  public resourceService: ResourceService;
+  constructor() {
 
-  /**
-  * Constructor to create injected service(s) object
-  *
-  * Default method of DeleteComponent class
-
-  * @param {ResourceService} resourceService Reference of ResourceService
- */
-  constructor(resourceService: ResourceService, private router: Router) {
-    this.resourceService = resourceService;
   }
-
-  ngOnInit() {
-  }
-  goToCreate() {
-    this.router.navigate(['/workspace/content/create']);
-  }
-
 }
