+++ src/app/migration/src/src/app/modules/workspace/components/content-creation-forms/create-textbook/create-textbook.component.ts
@@ -1,33 +1,15 @@
-import { Component, OnInit } from '@angular/core';
-import { ResourceService } from '@sunbird/shared';
-import { Router } from '@angular/router';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-create-textbook',
   templateUrl: './create-textbook.component.html',
   styleUrls: ['./create-textbook.component.css']
 })
-export class CreateTextbookComponent implements OnInit {
+export class CreateTextbookComponent {
 
-  /**
-   * To call resource service which helps to use language constant
-   */
-  public resourceService: ResourceService;
 
-  /**
-  * Constructor to create injected service(s) object
-  *
-  * Default method of DeleteComponent class
+  constructor() {
 
-  * @param {ResourceService} resourceService Reference of ResourceService
- */
-  constructor(resourceService: ResourceService, private router: Router) {
-    this.resourceService = resourceService;
   }
 
-  ngOnInit() {
-  }
-  goToCreate() {
-    this.router.navigate(['/workspace/content/create']);
-  }
 }
