+++ src/app/migration/src/src/app/modules/workspace/components/content-creation-forms/create-upload-content/create-upload-content.component.ts
@@ -1,33 +1,13 @@
-import { Component, OnInit } from '@angular/core';
-import { ResourceService } from '@sunbird/shared';
-import { Router } from '@angular/router';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-create-upload-content',
   templateUrl: './create-upload-content.component.html',
   styleUrls: ['./create-upload-content.component.css']
 })
-export class CreateUploadContentComponent implements OnInit {
+export class CreateUploadContentComponent {
 
-  /**
-   * To call resource service which helps to use language constant
-   */
-  public resourceService: ResourceService;
-
-  /**
-  * Constructor to create injected service(s) object
-  *
-  * Default method of DeleteComponent class
-
-  * @param {ResourceService} resourceService Reference of ResourceService
- */
-  constructor(resourceService: ResourceService, private router: Router) {
-    this.resourceService = resourceService;
+  constructor() {
   }
 
-  ngOnInit() {
-  }
-  goToCreate() {
-    this.router.navigate(['/workspace/content/create']);
-  }
 }
