+++ src/app/migration/src/src/app/modules/workspace/components/create-content/create-content.component.ts
@@ -1,5 +1,5 @@
 import { Router } from '@angular/router';
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 import { ResourceService } from '@sunbird/shared';
 import { SuiModule } from 'ng2-semantic-ui/dist';
 import { SuiModalService, TemplateModalConfig, ModalTemplate } from 'ng2-semantic-ui';
@@ -9,7 +9,7 @@ import { SuiModalService, TemplateModalConfig, ModalTemplate } from 'ng2-semanti
   templateUrl: './create-content.component.html',
   styleUrls: ['./create-content.component.css']
 })
-export class CreateContentComponent implements OnInit {
+export class CreateContentComponent {
 
   /**
    * To call resource service which helps to use language constant
@@ -27,6 +27,5 @@ export class CreateContentComponent implements OnInit {
   constructor(resourceService: ResourceService) {
     this.resourceService = resourceService;
   }
-  ngOnInit() {
-  }
+
  }
