+++ src/app/migration/src/src/app/modules/workspace/services/editor.service.spec.data.ts
@@ -1,0 +1,19 @@
+export const mockRes = {
+    createCollectionData: {
+        'id': 'api.content.create',
+        'ver': '1.0',
+        'ts': '2018-03-29T10:59:03.398Z',
+        'params': {
+           'resmsgid': '315e3060-3340-11e8-b827-71ac9e68d601',
+           'msgid': '31522270-3340-11e8-a6e1-b781c42eeb65',
+           'status': 'successful',
+           'err': null,
+           'errmsg': null
+        },
+        'responseCode': 'OK',
+        'result': {
+           'content_id': 'do_2124708548063559681134',
+           'versionKey': '1522321143354'
+        }
+     }
+};
